# docs/swagger.yaml
definitions:
  audit.Create:
    properties:
      app_id:
        type: string
      app_type:
        description: 0-企业  1-机构
        type: integer
      form_data:
        description: 审核表单信息 json
        type: string
    type: object
  audit.Search:
    properties:
      app_id:
        description: 企业id精确查询
        type: string
      app_name:
        description: 企业名字模糊查询
        type: string
      audit_ids:
        description: 审核id
        type: string
      page:
        type: integer
      page_size:
        type: integer
      registration_number:
        description: 注册号
        type: string
      states:
        description: 状态
        type: string
    type: object
  model.EnterpriseMuttable:
    properties:
      business_scope:
        type: string
      company_type:
        type: integer
      district:
        type: string
      estimate_value:
        type: integer
      industry:
        type: string
      introduction:
        type: string
      legal_representative:
        type: string
      legal_representative_id_img:
        type: string
      license_img:
        type: string
      name:
        type: string
      register_capital:
        type: integer
      registration_address:
        type: string
      registration_number:
        type: string
      shar_holders_json:
        type: string
      share_holder_info:
        type: string
      stage:
        type: integer
    type: object
  model.Valuate:
    properties:
      app_id:
        type: string
      created_at:
        description: 返回给业务侧
        type: string
      form_data:
        type: string
      requested_at:
        type: string
      state:
        type: integer
      udated_at:
        type: string
      valuate_id:
        type: string
    type: object
  request.Create:
    properties:
      data:
        $ref: '#/definitions/model.EnterpriseMuttable'
        description: 字段
      parent_id:
        description: 机构id 非必填
        type: string
      uid:
        description: 用户id
        type: string
    type: object
info:
  contact: {}
paths:
  /audit:
    get:
      description: 审核搜索
      parameters:
      - description: 字段注解
        in: body
        name: xxx
        schema:
          $ref: '#/definitions/audit.Search'
      produces:
      - application/json
      responses: {}
      summary: 审核搜索
      tags:
      - 审核
    post:
      description: 审核提交
      parameters:
      - description: 字段注解
        in: body
        name: xxx
        schema:
          $ref: '#/definitions/audit.Create'
      produces:
      - application/json
      responses: {}
      summary: 审核提交
      tags:
      - 审核
  /enterprise:
    post:
      description: 企业新建
      parameters:
      - description: 字段注解
        in: body
        name: xxx
        schema:
          $ref: '#/definitions/request.Create'
      produces:
      - application/json
      responses: {}
      summary: 企业新建
      tags:
      - 企业
  /valuate/search:
    get:
      description: 估值搜索
      parameters:
      - description: 企业ID
        in: query
        name: appID
        type: string
      - description: 页
        in: query
        name: page
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Valuate'
            type: array
      summary: 估值搜索
      tags:
      - 估值
swagger: "2.0"